+++ packages/sdk/src/messaging/thin-client.ts
@@ -201,6 +201,9 @@ function pickTelemetryOverrides(options: RelaycastTelemetryOptions): RelaycastTe
     ...(options.agentRelayUserId === undefined
       ? {}
       : { agentRelayUserId: options.agentRelayUserId }),
+    ...(options.agentRelayMachineId === undefined
+      ? {}
+      : { agentRelayMachineId: options.agentRelayMachineId }),
     ...(options.agentRelayOrgId === undefined ? {} : { agentRelayOrgId: options.agentRelayOrgId }),
     ...(options.agentRelayOrgSlug === undefined
       ? {}